Split content into different modules and trying to create basic implementations of all traits.
