+++ .vscode/settings.json
@@ -1,3 +1,4 @@
 {
     "files.eol": "\n",
+    "rust-analyzer.showUnlinkedFileNotification": false,
 }
