+++ src/base/mod.rs
@@ -1,0 +1,5 @@
+//!mod.rs (base)
+
+pub mod basenode;
+pub mod baseedge;
+pub mod basehead;
